Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1

SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

langchain==0.3.13
langchain-google-genai==2.0.7
pydantic==2.10.3
